+++ src/entities/profile/ui/profile-photos/profile-photos.tsx
@@ -1,7 +1,5 @@
 'use client'
 
-import { memo } from 'react'
-
 import { EMPTY_STRING } from '@/shared'
 import Image from 'next/image'
