+++ src/shared/ui/app-link/index.ts
@@ -1,0 +1,1 @@
+export { AppLink } from './app-link'
