+++ src/shared/ui/index.ts
@@ -24,3 +24,4 @@ export { TabContent, TabSwitcher } from './tabs'
 export { Text } from './text'
 export { ControlledTextarea, Textarea } from './textarea'
 export { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from './tooltip'
+export { AppLink } from './app-link'
